moved stuff to air_rs
